refactor(api): move OrganizationsToAttachToTargetProfile models in bounded context
