+++ api/tests/prescription/target-profile/unit/domain/models/OrganizationsToAttachToTargetProfile_test.js
@@ -1,4 +1,4 @@
-import { catchErr, domainBuilder, expect } from '../../../test-helper.js';
+import { catchErr, domainBuilder, expect } from '../../../../../test-helper.js';
 
 const id = 1;
 describe('Unit | Domain | Models | OrganizationsToAttachToTargetProfile', function () {
