+++ api/tests/tooling/domain-builder/factory/build-organizations-to-attach-to-target-profile.js
@@ -1,4 +1,4 @@
-import { OrganizationsToAttachToTargetProfile } from '../../../../lib/domain/models/OrganizationsToAttachToTargetProfile.js';
+import { OrganizationsToAttachToTargetProfile } from '../../../../src/prescription/target-profile/domain/models/OrganizationsToAttachToTargetProfile.js';
 
 const buildOrganizationsToAttachToTargetProfile = function ({ id = 123 } = {}) {
   return new OrganizationsToAttachToTargetProfile({
